remove recent addition of receive.denyNonFastForwards to walkthrough
That is a dangerous setting, and not needed.

I tried the walkthrough as it was without it, and syncing in the USB
remote worked ok. There was a problem merging from origin/master, since
that branch didn't exist yet, but that didn't prevent sync from working
at all.

To avoid even that problem, reordered the walkthrough, so files get
committed to the repo before the remote gets set up.

This commit was sponsored by Ole-Morten Duesund on Patreon.