remove stealth spam
While the link seems a valid article on recovering a sqlite database,
it's a stealth advertisement for a commercial product. Every post on
that blog is such a stealth advertisement.

Also, the question in this forum post has already been adequatly
answered.

So, I suspect this was a spamming attempt.